Broadcaster now receives serial input. Fixed Matlab code.
Broadcaster now only transmits once it's received an update to the
current state of the game. There is an issue currently where the
receiver is getting the transmission shifted out by 1 byte.

E.g. Sending "test", the receiver is getting "est". Observing both sets
of code this shouldn't be happening. This must be fixed.

Matlab changes:
Open and close serial file locally, initial values to game struct for
testing, fixed bugs in defining A array.