Fixed broadcaster issues with serial.
Broadcaster can now work well with serial, by introducing a slight delay
in Serial read, and simply increasing the amount of retransmissions
performed to ensure the packet is received.

Note: The broadcaster seems to mess up the first few packets, but after
that it's fine. For example, you can easily sent 100 different game
states and lose the first 5 while the rest are perfect.

I've also increased the speed with which we can send data from matlab,
though this is directly dependent on radio parameters and an external
parameter that I've set, NUM_RETRANS.